Configuration

The following options are currently supported to be set via base::options().

mlrng.verbose

Verbosity. Set to FALSE to suppress some output.

mlrng.debug

Debug mode. Set to TRUE to enable additional output and some (slower) checks.

mlrng.keep.train.output

Store the output of the models in the log file. Default: FALSE.

mlrng.continue.on.learner.error

Keep running if a learner encounters a problem. The resulting model will be a model fitted by a dummy learner.

mlrng.train.encapsulation

Control how the training of learners is encapsulated. “none” means no encapsulation but also no logging (fastest). If set to “evaluate” (default), the package evaluate is used to run the model fit. If set to “callr”, the package callr starts the training in a separate R process to guard your session against segfaults.

These options may be set in a configuration file which is automatically parsed on package load. To get the location according to your operating system, see the example. readMlrngConfig() reads a configuration file and returns its settings in a named list, writeMlrngConfig() writes a configuration file with all currently set options.

readMlrngConfig(conf = file.path(rappdirs::user_config_dir("mlrng"),
  "config.yml"))

writeMlrngConfig(conf = file.path(rappdirs::user_config_dir("mlrng"),
  "config.yml"))

Arguments

conf
(character(1))
Location of the configuration file to load or write.

Examples

# Location of the config file for your system:
conf = file.path(rappdirs::user_config_dir("mlrng"), "config.yml")
print(conf)

#> [1] "/home/travis/.config/mlrng/config.yml"

# Current settings:
print(readMlrngConfig())

#> list()